Finish

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jabae <jabae@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/21 13:26:03 by jabae             #+#    #+#             */
-/*   Updated: 2022/03/24 17:46:06 by jabae            ###   ########.fr       */
+/*   Updated: 2022/03/24 18:17:54 by jabae            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,6 @@
 
 #include <unistd.h>
 #include <stdarg.h>
-#include <stdlib.h>
 
 int	ft_printf(const char *format, ...);
 int	ft_switch_format(va_list ap, const char *format, int *idx);
@@ -27,7 +26,7 @@ char	*ft_make_hex(char c);
 int		ft_putstr_count(char *str);
 int		ft_putchar_count(char c);
 int		ft_putint_count(long long num);
-int		ft_putunsignedint_count(unsigned long long num);
+int		ft_putunsignedint_count(unsigned int num);
 int		ft_puthex_count(unsigned int num, char *hexbase);
 int		ft_putpointer_count(unsigned long long num, char *hexbase);
